+++ intelmq/lib/pipeline.py
@@ -62,7 +62,6 @@ class Pipeline(object):
             elif type_ is str:
                 q = {"_default": queues.split()}
             elif type_ is dict:
-                # if "_default" not in queues: raise KeyError("Missing _default key.")
                 q = queues
                 for key, val in queues.items():
                     q[key] = val if type(val) is list else val.split()
@@ -119,16 +118,16 @@ class Redis(Pipeline):
         self.load_configurations(queues_type)
         super(Redis, self).set_queues(queues, queues_type)
 
-    def send(self, message, queue="_default"):
+    def send(self, message, path="_default"):
         message = utils.encode(message)
         try:
-            queues = self.destination_queues[queue]
+            queues = self.destination_queues[path]
         except KeyError as exc:
             raise exceptions.PipelineError(exc)
         if self.load_balance:
             queues = [queues[self.load_balance_iterator]]
             self.load_balance_iterator += 1
-            if self.load_balance_iterator == len(self.destination_queues[queue]):
+            if self.load_balance_iterator == len(self.destination_queues[path]):
                 self.load_balance_iterator = 0
 
         for destination_queue in queues:
@@ -213,9 +212,9 @@ class Pythonlist(Pipeline):
         for destination_queue in chain.from_iterable(self.destination_queues.values()):
             self.state[destination_queue] = []
 
-    def send(self, message, queue="_default"):
+    def send(self, message, path="_default"):
         """Sends a message to the destination queues"""
-        for destination_queue in self.destination_queues[queue]:
+        for destination_queue in self.destination_queues[path]:
             if destination_queue in self.state:
                 self.state[destination_queue].append(utils.encode(message))
             else: