+++ CHANGELOG.md
@@ -32,6 +32,7 @@ Support for Python 3.3 has been dropped, it reached its end of life.
   These functions will be called by `intelmqctl check` if the bot is configured with the given parameters
 - Add `RewindableFileHandle` to utils making handling of CSV files more easy (optionally)
 - lib/bot: top level bot parameters (description, group, module, name) are exposed as members of the class.
+- you may now define more than one destination queues path the bot should pass the message to, see [Pipelines](https://github.com/certtools/intelmq/blob/develop/docs/User-Guide.md#pipeline-configuration)
 
 ### Bots
 #### Collectors